feature(api): update get organization admin to retrieve feature flag
