+++ api/lib/application/organizations/organization-controller.js
@@ -36,13 +36,6 @@ const { mapCertificabilityByLabel } = require('./helpers.js');
 const csvSerializer = require('../../infrastructure/serializers/csv/csv-serializer.js');
 
 module.exports = {
-  async getOrganizationDetails(request, h, dependencies = { organizationForAdminSerializer }) {
-    const organizationId = request.params.id;
-
-    const organizationDetails = await usecases.getOrganizationDetails({ organizationId });
-    return dependencies.organizationForAdminSerializer.serialize(organizationDetails);
-  },
-
   async create(request) {
     const superAdminUserId = extractUserIdFromRequest(request);
     const organization = organizationForAdminSerializer.deserialize(request.payload);
