+++ api/tests/unit/application/organizations-administration/organization-administration-controller_test.js
@@ -77,4 +77,31 @@ describe('Unit | Application | Organizations | organization-administration-contr
       expect(response.source).to.deep.equal(serializedOrganization);
     });
   });
+
+  describe('#getOrganizationDetails', function () {
+    it('should call the usecase and serialize the response', async function () {
+      // given
+      const organizationId = 1234;
+      const request = { params: { id: organizationId } };
+
+      const organizationDetails = Symbol('organizationDetails');
+      const organizationDetailsSerialized = Symbol('organizationDetailsSerialized');
+      sinon.stub(usecases, 'getOrganizationDetails').withArgs({ organizationId }).resolves(organizationDetails);
+      const organizationForAdminSerializerStub = {
+        serialize: sinon.stub(),
+      };
+
+      organizationForAdminSerializerStub.serialize.withArgs(organizationDetails).returns(organizationDetailsSerialized);
+
+      const dependencies = {
+        organizationForAdminSerializer: organizationForAdminSerializerStub,
+      };
+
+      // when
+      const result = await organizationController.getOrganizationDetails(request, hFake, dependencies);
+
+      // then
+      expect(result).to.equal(organizationDetailsSerialized);
+    });
+  });
 });
